+++ addon/token.ts
@@ -204,7 +204,7 @@ const expand = (
 					prev,
 				);
 			} else if (Parser.functionHooks.has(name)) {
-				return Parser.functionHooks.get(name)!(target, context || undefined);
+				return context === false ? m : Parser.functionHooks.get(name)!(target, context || undefined);
 			} else if (expandedMagicWords.has(name)) {
 				if (context === false) {
 					return m;